refactor: Move `isUUID()` to string utils
