+++ lib/modules/platform/bitbucket/index.ts
@@ -8,8 +8,9 @@ import * as git from '../../../util/git';
 import * as hostRules from '../../../util/host-rules';
 import { BitbucketHttp, setBaseUrl } from '../../../util/http/bitbucket';
 import type { HttpOptions } from '../../../util/http/types';
-import { isUUID, regEx } from '../../../util/regex';
+import { regEx } from '../../../util/regex';
 import { sanitize } from '../../../util/sanitize';
+import { isUUID } from '../../../util/string';
 import type {
   BranchStatusConfig,
   CreatePRConfig,
