+++ lib/util/string.ts
@@ -117,3 +117,11 @@ export function anyMatchRegexOrMinimatch(
 ): boolean | null {
   return patterns.some((pattern) => matchRegexOrMinimatch(input, pattern));
 }
+
+const UUIDRegex = regEx(
+  /^\{[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}\}$/i,
+);
+
+export function isUUID(input: string): boolean {
+  return UUIDRegex.test(input);
+}
